Major refactor to resolve namesnipe
