+++ CONTRIBUTING.md
@@ -1,9 +1,9 @@
-# Welcome to the PyminderContributor Guidelines!
+# Welcome to the PiminderContributor Guidelines!
 Hey, thanks for checking us out!
 
 First off, let me just say that any contribution matters - from pull-requesting in new functions, to requesting new features, or even just reporting your crashes and bugs. Clearly, there's potentially a lot of work to be done here if this becomes and ongoing-development project.
  
-Pyminder doesn't have automated crash and bug reporting, and its unlikely it ever will. I'm just one guy with a few spare hours in the evening trying to put this thing together. Everything you contribute helps.
+Piminder doesn't have automated crash and bug reporting, and its unlikely it ever will. I'm just one guy with a few spare hours in the evening trying to put this thing together. Everything you contribute helps.
 
 Secondly, we do have a few resources for the eager:
 - [**Discord Server:**](https://discord.gg/34KJfkg) easily the fastest way to join in the conversation or ask a question;
@@ -25,14 +25,14 @@ If you have a feature request for Tapestry, go ahead and open a new issue at thi
 We kind of can't overstate the importance of providing your bug reports to us. It's so important we've made a template for them. When Tapestry crashes on your system, we have no idea what happened or why. Please be encouraged to file an issue and fill out the bugs form. Super important stuff.
 
 # Contributing with Code: The Pull Request!
-For Pyminder, we use Fork and Request. If you want to contribute to the development process with your code, the preferred method is to first fork the repo, make your changes there, and then submit a pull request. For obvious reasons all submissions are subject to review. To make this whole process easier, be sure to use nice, clear, descriptive code. Provide documentation and comments as needed.
+For Piminder, we use Fork and Request. If you want to contribute to the development process with your code, the preferred method is to first fork the repo, make your changes there, and then submit a pull request. For obvious reasons all submissions are subject to review. To make this whole process easier, be sure to use nice, clear, descriptive code. Provide documentation and comments as needed.
 
 In general though, we're super happy to have any help that we can. People who contribute with code are being kept track of and will be included in the credits.
 
 # Contributing with Coins
-Pyminder is and ever will be free-as-in-beer software, provided at no charge and by donation to users of all scales.
+Piminder is and ever will be free-as-in-beer software, provided at no charge and by donation to users of all scales.
 
-As a freeware project, we're obviously working with the bare minimum funding possible - luckily, this sort of thing also doesn't take much. If you'd like to donate to keep Pyminder's development alive, a one-time donation can be made by purchasing us some ~~debugger fluid~~ coffee via Ko-Fi.com:
+As a freeware project, we're obviously working with the bare minimum funding possible - luckily, this sort of thing also doesn't take much. If you'd like to donate to keep Piminder's development alive, a one-time donation can be made by purchasing us some ~~debugger fluid~~ coffee via Ko-Fi.com:
 
 [![ko-fi](https://www.ko-fi.com/img/githubbutton_sm.svg)](https://ko-fi.com/A8614WZ3)
