+++ src/piminder_monitor/monitor.conf
@@ -1,4 +1,4 @@
-[Pyminder Monitor Config]
+[Piminder Monitor Config]
 color_resting: #53878a
 major_error_color: #FF0000
 minor_error_color: #FFF70F
