+++ src/piminder_service/pyminder-service.conf
@@ -5,7 +5,7 @@ DEBUG: True # Controls if the debug option in
 [Network Configuration]
 LISTENHOST: 0.0.0.0
 LISTENPORT: 80
-NETWORK_LABEL: Pyminder One-Day Build Test Instance
+NETWORK_LABEL: Piminder One-Day Build Test Instance
 
 USE_SSL=False  # Can be false if the docker container is part of an reverse proxy cluster!
 SSL_CERT=cert.pem
