+++ src/piminder_service/resources/users.py
@@ -1,26 +1,26 @@
 """
-This script is a component of pyminder's back-end controller.
-This resource handles the registration of messages and handling existing messages through various restful methods.
+This script is a component of Piminder's back-end controller.
+This resource handles the registration, deletion, and modification of users and relies on an authorization
+confirmation method from utilities.py.
 
 Author: Zac Adam-MacEwen (zadammac@kenshosec.com)
 An Arcana Labs utility.
 
 Produced under license.
 Full license and documentation to be found at:
-https://github.com/ZAdamMac/Enumpi-C2
+https://github.com/ZAdamMac/Piminder
 """
 
-import datetime
+import bcrypt
 from flask_restful import Resource
 from flask import current_app, request, make_response
 import pymysql
-import uuid
 from .utilities import authenticated_exec, basic_auth, json_validate
 
-__version__ = "0.2.0"
+__version__ = "prototype"
 
 
-class MessageAPI(Resource):
+class UsersAPI(Resource):
     def get(self):
         """An authenticated user with reporting permissions may retrieve a
         full listing of all commands logged by the server set up on the server.
@@ -32,7 +32,7 @@ class MessageAPI(Resource):
             connection = pymysql.connect(host=current_app.config["DBHOST"],
                                          user=current_app.config["USERNAME"],
                                          password=current_app.config["PASSPHRASE"],
-                                         db='pyminder',
+                                         db='Piminder',
                                          cursorclass=pymysql.cursors.DictCursor)
             connection.ping(reconnect=True)
         except KeyError:
@@ -41,7 +41,7 @@ class MessageAPI(Resource):
             return {'message': 'Internal Server Error'}, 500
         proceed, user = basic_auth(cookie, connection)
         if proceed:  # A chicken ain't nothing but a bird.
-            dict_return = authenticated_exec(user, 2, connection, messages_get, "")
+            dict_return = authenticated_exec(user, 3, connection, users_get, "")
             resp = make_response(dict_return)
             resp.status_code = dict_return["error"]
             resp.content_type = "application/json"
@@ -59,7 +59,7 @@ class MessageAPI(Resource):
             connection = pymysql.connect(host=current_app.config["DBHOST"],
                                          user=current_app.config["USERNAME"],
                                          password=current_app.config["PASSPHRASE"],
-                                         db='pyminder',
+                                         db='Piminder',
                                          cursorclass=pymysql.cursors.DictCursor)
             connection.ping(reconnect=True)
         except KeyError:
@@ -68,7 +68,7 @@ class MessageAPI(Resource):
             return {'message': 'Internal Server Error, sql error'}, 501
         proceed, user = basic_auth(cookie, connection)
         if proceed:  # A chicken ain't nothing but a bird.
-            dict_return = authenticated_exec(user, 1, connection, messages_post, request.get_json())
+            dict_return = authenticated_exec(user, 3, connection, users_post, request.get_json())
             resp = make_response(dict_return)
             resp.status_code = dict_return["error"]
             resp.content_type = "application/json"
@@ -82,13 +82,12 @@ class MessageAPI(Resource):
 
         :return:
         """
-
         cookie = request.headers.get("Authorization")
         try:
             connection = pymysql.connect(host=current_app.config["DBHOST"],
                                          user=current_app.config["USERNAME"],
                                          password=current_app.config["PASSPHRASE"],
-                                         db='pyminder',
+                                         db='Piminder',
 
                                          cursorclass=pymysql.cursors.DictCursor)
             connection.ping(reconnect=True)
@@ -98,7 +97,7 @@ class MessageAPI(Resource):
             return {'message': 'Internal Server Error'}, 500
         proceed, user = basic_auth(cookie, connection)
         if proceed:  # A chicken ain't nothing but a bird.
-            dict_return = authenticated_exec(user, 2, connection, messages_patch, request.get_json())
+            dict_return = authenticated_exec(user, 3, connection, users_patch, request.get_json())
             resp = make_response(dict_return)
             resp.status_code = dict_return["error"]
             resp.content_type = "application/json"
@@ -111,13 +110,12 @@ class MessageAPI(Resource):
 
         :return:
         """
-
         cookie = request.headers.get("Authorization")
         try:
             connection = pymysql.connect(host=current_app.config["DBHOST"],
                                          user=current_app.config["USERNAME"],
                                          password=current_app.config["PASSPHRASE"],
-                                         db='pyminder',
+                                         db='Piminder',
 
                                          cursorclass=pymysql.cursors.DictCursor)
             connection.ping(reconnect=True)
@@ -127,7 +125,7 @@ class MessageAPI(Resource):
             return {'message': 'Internal Server Error'}, 500
         proceed, user = basic_auth(cookie, connection)
         if proceed:  # A chicken ain't nothing but a bird.
-            dict_return = authenticated_exec(user, 2, connection, messages_delete, request.get_json())
+            dict_return = authenticated_exec(user, 3, connection, users_delete, request.get_json())
             resp = make_response(dict_return)
             resp.status_code = dict_return["error"]
             resp.content_type = "application/json"
@@ -138,60 +136,54 @@ class MessageAPI(Resource):
 # Here follow the actual actions!
 
 
-def messages_get(body, connection):
-    """A stored join function that gets all the currently registered commands,
-    their relevant metadata, the name of the client they are associated with
-    and the message, if any. This is returned to the requestor in a JSON
-    format for further processing."""
+def users_get(discard, connection):
+    """A stored join function that gets all the currently registered users, along with permission level"""
+    del discard
     cur = connection.cursor()
-    del body
+
     # Unwieldy command handles most of the data processing in SQL which is faster than doing this in python.
 
-    cmd = "SELECT * FROM messages ORDER BY time_raised desc;"
+    cmd = "SELECT username, memo, permlevel FROM users ORDER BY username asc;"  # we don't want the passwords!
     cur.execute(cmd)
     messages = cur.fetchall()
     response = {}
     counter = -1
     for message in messages:
-        output_keys = {"id": "messageId", "name": "name", "read_flag": "read", "errorlevel": "errorLevel",
-                       "time_raised": "timestamp", "message": "message"}
+        output_keys = {"username": "username", "memo": "memo", "permlevel": "permissionLevel"}
         counter += 1
         this_message = {}
         for key in output_keys.keys():
             this_message.update({output_keys[key]: message[key]})
-        if this_message["read"] == b'\x00':
-            this_message.update({"read": False})
-        else:
-            this_message.update({"read": True})
-        time_out = this_message["timestamp"].strftime("%Y-%m-%dT%H:%M:%SZ")  # It is nice to have ISO 8601 compliance
-        this_message.update({"timestamp": time_out})
         response.update({str(counter): this_message})
     response.update({"error": 200})
 
     return response
 
 
-def messages_post(body, connection):
-    """A very simplistic function that adds a fresh command to the commands table."""
+def users_post(body, connection):
+    """A very simplistic function that adds a fresh user to the users table."""
     cur = connection.cursor()
-    dict_schema = {"name": "", "timestamp": "", "errorlevel": "", "message": ""}
+    dict_schema = {"username": "", "permissionLevel": "", "memo": "", "password": ""}
     json_valid, errors = json_validate(body, dict_schema)
-    if body["errorlevel"] not in ["info", "minor", "major"]:
+    dict_levels = {"service": 1, "monitor": 2, "admin": 3}  # This dictionary defines all available levels. TODO DOC
+    if body["permissionLevel"] not in dict_levels.keys():
         json_valid = False
-        errors.update({"errorlevel": "Error level not one of info, minor, or major."})
+        errors.update({"errorlevel": "Error level not one of service, monitor, or admin."})
+    else:
+        body["permlevel"] = dict_levels[body["permissionLevel"]]  # Replace friendly string with level integer
+
+    password = body["password"].encode('utf8')  # Bcrypt operates on byte arrays, not strings
+    stored_password = bcrypt.hashpw(password, bcrypt.gensalt())  # I prefer my hash as salty as practical
+    body["password"] = stored_password.decode('utf8')  # And then we drop it back to a string to run.
 
     if json_valid:
         d_message = {}
         d_message.update(body)
-        d_message.update({"id": str(uuid.uuid4())})
-        d_message.update({"read": False})
-        d_message.update({"timestamp": datetime.datetime.strptime(body["timestamp"], "%Y-%m-%dT%H:%M:%SZ").timestamp()})
-        cmd = "INSERT INTO messages " \
-              "(id, name, time_raised, errorlevel, message, read_flag) " \
-              "VALUES (%(id)s, %(name)s, FROM_UNIXTIME(%(timestamp)s), %(errorlevel)s, " \
-              "%(message)s, %(read)s)"
+        cmd = "INSERT INTO users " \
+              "(username, password, memo, permlevel) " \
+              "VALUES (%(username)s, %(password)s, %(memo)s, %(permlevel)s);"
         cur.execute(cmd, d_message)
-        response = {"error": 200}
+        response = {"error": 200, "message": ("User %s created successfully." % d_message["username"])}
         connection.commit()
     else:
         response = {"all_errors": errors, "error": 400}
@@ -199,93 +191,59 @@ def messages_post(body, connection):
     return response
 
 
-def messages_patch(body, connection):
-    """This function indicates in the DB that a message has been read.
-    A stored procedure or cron job will garbage collect."""
+def users_patch(body, connection):
+    """This function updates an existing user and should only be callable by an admin. Mostly used to change PW"""
     cur = connection.cursor()
-    dict_schema = {"messageId": ""}
+    dict_schema = {"username": "", "permissionLevel": "", "memo": "", "password": ""}
     json_valid, errors = json_validate(body, dict_schema)
+    dict_levels = {"service": 1, "monitor": 2, "admin": 3}  # This dictionary defines all available levels.
+    if body["permissionLevel"] not in dict_levels.keys():
+        json_valid = False
+        errors.update({"errorlevel": "Error level not one of service, monitor, or admin."})
+    else:
+        body["permissionLevel"] = dict_levels[body["permissionLevel"]]  # Replace friendly string with level integer
+
+    password = body["password"].encode('utf8')  # Bcrypt operates on byte arrays, not strings
+    stored_password = bcrypt.hashpw(password, bcrypt.gensalt())  # I prefer my hash as salty as practical
+    body["password"] = stored_password.decode('utf8')  # And then we drop it back to a string to run.
 
     if json_valid:
-        # First, make sure this hasn't already been executed.
-        cmd = "SELECT id, read_flag " \
-              "FROM messages " \
-              "WHERE id=%s"
-        cur.execute(cmd, body["messageId"])
-        response = cur.fetchone()
-        print(response)
-        # Gotta be a better way to do the following line, but...
-        try:
-            if response["read_flag"] == b'\x00':
-                cmd = "UPDATE messages " \
-                      "SET read_flag=TRUE " \
-                      "WHERE id=%s"
-                cur.execute(cmd, body["messageId"])
-                response = {"error": 200}
-                connection.commit()
-            else:
-                response = {"error": 400, "message": "Could not delete command - already acknowledged."}
-        except KeyError:
-            response = {"error": 400, "message": "The indicated ID does not exist in the messages table."}
+        d_message = {}
+        d_message.update(body)
+        cmd = "UPDATE users " \
+              "SET password=%(password)s, memo=%(memo)s, permlevel=%(permissionLevel)s " \
+              "WHERE username like %(username)s;"
+        cur.execute(cmd, d_message)
+        response = {"error": 200, "message": ("User %s updated successfully." % d_message["username"])}
+        connection.commit()
     else:
         response = {"all_errors": errors, "error": 400}
 
     return response
 
 
-def messages_patch(body, connection):
-    """This function indicates in the DB that a message has been read.
-    A stored procedure or cron job will garbage collect."""
+def users_delete(body, connection):
+    """This function removes the selected user's permissions to do anything while leaving them in the DB."""
     cur = connection.cursor()
-    dict_schema = {"messageId": ""}
+    dict_schema = {"username": ""}
     json_valid, errors = json_validate(body, dict_schema)
 
     if json_valid:
         # First, make sure this hasn't already been executed.
-        cmd = "SELECT id, read_flag FROM messages WHERE id=%s"
-        cur.execute(cmd, body["messageId"])
-        response = cur.fetchone()
-        # Gotta be a better way to do the following line, but...
-        try:
-            if response["read_flag"] == b'\x00':
-                cmd = "UPDATE messages " \
-                      "SET read_flag=TRUE " \
-                      "WHERE id=%s"
-                cur.execute(cmd, body["messageId"])
-                response = {"error": 200}
-                connection.commit()
-            else:
-                response = {"error": 400, "message": "Could not mark message as read; already read."}
-        except KeyError:
-            response = {"error": 400, "message": "The indicated ID does not exist in the messages table."}
-    else:
-        response = {"all_errors": errors, "error": 400}
-
-    return response
-
-
-def messages_delete(body, connection):
-    """This function removes the selected entity from database prior to garbage collection."""
-    cur = connection.cursor()
-    dict_schema = {"messageId": ""}
-    json_valid, errors = json_validate(body, dict_schema)
-
-    if json_valid:
-        # First, make sure this hasn't already been executed.
-        cmd = "SELECT id FROM messages WHERE id=%s"
-        cur.execute(cmd, body["messageId"])
+        cmd = "SELECT username FROM users WHERE username like %(username)s;"
+        cur.execute(cmd, body)
         response = cur.fetchone()
         # Gotta be a better way to do the following line, but...
         try:
             if response:
-                cmd = "DELETE FROM messages WHERE id=%s"
-                cur.execute(cmd, body["messageId"])
+                cmd = "UPDATE users SET permlevel=0 WHERE username like %(username)s;"
+                cur.execute(cmd, body)
                 response = {"error": 200}
                 connection.commit()
             else:
-                response = {"error": 400, "message": "Could not delete message, already deleted?"}
+                response = {"error": 400, "message": "Could not deactivate user"}
         except KeyError:
-            response = {"error": 400, "message": "The indicated ID does not exist in the messages table."}
+            response = {"error": 400, "message": "The indicated user does not exist"}
     else:
         response = {"all_errors": errors, "error": 400}
